Fix bug when handling cycles.
Cycles were not being handled correctly and could result in a
runtime panic because RefCells were being recursively and mutably
borrowed (can't borrow a RefCell mutably more than once).

This is solved by moving the garbace collection header into a
separate cell from the object contents, and using Cell for it
rather than RefCell.  Then the contents can be mutable without
the need for a mutable borrow.

Thanks to talchas in #rust on irc.mozilla.org for the help!